Migrate remainder of DefaultCliAdapter tests for pico CLI
